requirements.txt
alembic
Flask
Flask-Login
Flask-Mail
Flask-Migrate
Flask-SQLAlchemy
email_validator
gunicorn
numpy
opencv-python
openpyxl
pandas
pillow
psycopg2-binary
PyMuPDF
python-dotenv
reportlab
SQLAlchemy
tkcalendar
Werkzeug
